YeniÇalışma

+++ ETicaretAPI/Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -1,9 +1,4 @@
 ﻿using ETicaretAPI.Domain.Entities.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ETicaretAPI.Domain.Entities
 {
@@ -11,7 +6,7 @@ namespace ETicaretAPI.Domain.Entities
     {
         public string Name { get; set; }
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public float Price { get; set; }
         public ICollection<Order> Orders { get; set; }
 
     }
